requirements.txt
psutil==5.9.5
pyyaml==6.0
prometheus-client==0.17.0